netif: Create netif layer, initially provide datagram read/write
Add in skeleton coap_netif.c coap_netif_internal.h files, populated with
coap_netif_dgrm_read(), coap_netif_dgrm_read_ep(), and coap_netif_dgrm_write()
functions that are called by the CoAP and (D)TLS code which in turn call
the appropriate coap_socket_*() functions.

The coap_dtls_send() function parameters are aligned with
coap_netif_dgrm_write().

+++ include/coap3/coap_dtls_internal.h
@@ -192,9 +192,9 @@ void coap_dtls_session_update_mtu(coap_session_t *coap_session);
  * @return @c 0 if this would be blocking, @c -1 if there is an error or the
  *         number of cleartext bytes sent.
  */
-int coap_dtls_send(coap_session_t *coap_session,
-                   const uint8_t *data,
-                   size_t data_len);
+ssize_t coap_dtls_send(coap_session_t *coap_session,
+                       const uint8_t *data,
+                       size_t data_len);
 
 /**
  * Check if timeout is handled per CoAP session or per CoAP context.
